+++ packages/react-mui-store-file-viewer/src/components/CreateNewFileButton/index.tsx
@@ -5,6 +5,7 @@ import FileOpenIcon from "@mui/icons-material/FileOpen";
 import { useStoreFileViewerContext } from "../../hooks/context";
 import { ObjectStore } from "../../types";
 import { createMultipartUploadForm } from "../../utils/file";
+import { uploadFiles } from "../UploadPopover/hooks";
 
 interface Props {
   ButtonProps?: ButtonProps;
@@ -19,36 +20,9 @@ const CreateNewFileButton: FC<Props> = ({ ButtonProps, onClose }) => {
 
   const [loading, setLoading] = useState(false);
 
-  const obj = useMemo<ObjectStore.FolderCreateRequest>(() => {
-    return {
-      name: "",
-      path: "/",
-    };
-  }, []);
-
   const handleClose = () => {
     setOpen(false);
     onClose();
-  };
-
-  const handleCreateFile = async (
-    files: FileList,
-    body: ObjectStore.PutRequest
-  ) => {
-    if (!apiRef.upload || !files) return;
-
-    setLoading(true);
-    return apiRef
-      .upload(createMultipartUploadForm(files, body), "multipart/form-data")
-      .then(() => {
-        setLoading(false);
-        handleClose();
-        return reload();
-      })
-      .catch((error) => {
-        setLoading(false);
-        handleClose();
-      });
   };
 
   return (
@@ -60,12 +34,8 @@ const CreateNewFileButton: FC<Props> = ({ ButtonProps, onClose }) => {
         type="file"
         onChange={(e) => {
           if (e.target.files) {
-            handleCreateFile(e.target.files, {
-              name: e.target.files[0].name ?? "Unknown",
-              ...(!!parent && {
-                parent_id: parent.id,
-              }),
-            });
+            uploadFiles(e.target.files);
+            handleClose();
           }
         }}
       />