+++ packages/react-mui-store-file-viewer/src/components/FileDropBox/index.tsx
@@ -1,12 +1,24 @@
-import { Box, Stack, Typography, LinearProgress } from "@mui/material";
+import {
+  Box,
+  Stack,
+  Typography,
+  LinearProgress,
+  Skeleton,
+  Button,
+} from "@mui/material";
 import { SxProps, useTheme, lighten } from "@mui/material/styles";
+import find from "lodash/find";
 
-import React, { FC, useMemo, useRef, useState } from "react";
+import React, { FC, useEffect, useMemo, useRef, useState } from "react";
 import { ObjectStore } from "../../types";
 import { useStoreFileViewerContext } from "../../hooks/context";
 import { createMultipartUploadForm } from "../../utils/file";
 
+import AddIcon from "@mui/icons-material/Add";
+
 import FileUploadOutlinedIcon from "@mui/icons-material/FileUploadOutlined";
+import { uploadFiles, useUploadPopover } from "../UploadPopover/hooks";
+import TargetPopper from "./TargetPopper";
 
 interface Props {
   children?: React.ReactNode;
@@ -15,45 +27,26 @@ interface Props {
 
 const FileDropBox: FC<Props> = ({ children, BoxSx }) => {
   const theme = useTheme();
-  const { apiRef, reload, parent } = useStoreFileViewerContext();
+  const { parent, data, loading } = useStoreFileViewerContext();
+  const objects = useRef(data?.objects ?? []);
+  objects.current = data?.objects ?? [];
 
-  const isVisibleOnHoverOnly = !!children;
+  const [targetParent, setTargetParent] = useState(parent);
+  const parentRef = useRef(targetParent);
+  parentRef.current = targetParent;
+
+  useEffect(() => {
+    setTargetParent(parent);
+  }, [parent]);
+
+  const isMissingData = !loading && !parent && !data.objects.length;
 
   const drop = useRef<any>(null);
-  const drag = useRef<any>(null);
+  const exit = useRef<any>(null);
+
   const inputRef = useRef<HTMLInputElement | undefined>();
 
-  const [loading, setLoading] = useState(false);
   const [dragging, setDragging] = useState(false);
-
-  const body = useMemo<ObjectStore.PutRequest>(() => {
-    return {
-      name: "unknown",
-    };
-  }, []);
-
-  const handleCreateFile = async (files: FileList) => {
-    if (!apiRef.upload || !files) return;
-    setDragging(false);
-
-    setLoading(true);
-    return apiRef
-      .upload(
-        createMultipartUploadForm(files, {
-          ...body,
-          ...(!!parent && { parent_id: parent.id }),
-          name: files[0].name ?? body.name,
-        }),
-        "multipart/form-data"
-      )
-      .then(() => {
-        setLoading(false);
-        reload();
-      })
-      .catch((error) => {
-        setLoading(false);
-      });
-  };
 
   React.useEffect(() => {
     drop.current?.addEventListener("dragover", handleDragOver);
@@ -69,9 +62,46 @@ const FileDropBox: FC<Props> = ({ children, BoxSx }) => {
     };
   }, []);
 
+  const updateTarget = (element: any) => {
+    if (element.getAttribute("data-field") === "name") {
+      const row = element.parentElement;
+      if (row) {
+        const objectId = row.getAttribute("data-id");
+        if (objectId) {
+          const object = find(objects.current, (o) => o.id === objectId);
+          if (object && object.type === "folder") {
+            setTargetParent(object);
+            return true;
+          }
+        }
+      }
+    }
+
+    return false;
+  };
+
   const handleDragOver = (e: any) => {
     e.preventDefault();
     e.stopPropagation();
+
+    if (exit.current) {
+      clearTimeout(exit.current);
+      exit.current = null;
+    }
+    setDragging(true);
+
+    if (e.target.getAttribute("data-field") === "name") {
+      if (!updateTarget(e.target)) {
+        setTargetParent(parent);
+      }
+    } else {
+      const cell = e.target.closest(".MuiDataGrid-Cell");
+      if (cell && updateTarget(cell)) {
+        // Do nothing, parent is targetted
+      } else {
+        setTargetParent(parent);
+      }
+    }
   };
 
   const handleDrop = (e: any) => {
@@ -81,32 +111,31 @@ const FileDropBox: FC<Props> = ({ children, BoxSx }) => {
     const { files } = e.dataTransfer;
 
     if (files && files.length) {
-      handleCreateFile(files);
+      uploadFiles(files, parentRef.current);
     }
+    setDragging(false);
   };
 
   const handleDragEnter = (e: any) => {
     e.preventDefault();
     e.stopPropagation();
-
-    if (e.target !== drag.current) {
-      setDragging(true);
-    }
   };
 
   const handleDragLeave = (e: any) => {
     e.preventDefault();
     e.stopPropagation();
 
-    if (e.target === drag.current) {
-      setDragging(false);
+    if (!exit.current) {
+      exit.current = setTimeout(() => {
+        setDragging(false);
+      }, 200);
     }
   };
 
-  const main = theme.palette.secondary.dark;
+  const main = theme.palette.info.main;
   return (
     <>
-      {!isVisibleOnHoverOnly && (
+      {
         <input
           // @ts-ignore
           ref={inputRef}
@@ -114,86 +143,108 @@ const FileDropBox: FC<Props> = ({ children, BoxSx }) => {
           type="file"
           onChange={(e) => {
             if (e.target.files) {
-              handleCreateFile(e.target.files);
+              uploadFiles(e.target.files);
             }
           }}
         />
-      )}
-      <Box
+      }
+      <Stack
         ref={drop}
+        direction="row"
         sx={{
           position: "relative",
-          ...(!isVisibleOnHoverOnly && {
+          width: "100%",
+          border: `2px solid transparent`,
+          ...(!!dragging && {
             cursor: "pointer",
             border: `2px solid ${main}`,
-            borderStyle: "dashed",
-            width: "100%",
-            height: "150px",
           }),
-          ":hover": {
-            cursor: "pointer",
-            border: `2px dashed solid ${main}`,
-            borderStyle: "dashed",
-          },
           ...BoxSx,
         }}
-        onClick={(e) => {
-          if (isVisibleOnHoverOnly || loading) return;
-
-          e.stopPropagation();
-          e.preventDefault();
-          inputRef.current?.click();
-        }}
       >
-        {dragging && (
-          <Box
-            ref={drag}
-            sx={{
-              position: "absolute",
-              top: 0,
-              left: 0,
-              right: 0,
-              bottom: 0,
-              width: "100%",
-              height: "100%",
-            }}
-          ></Box>
-        )}
-        {children ?? (
-          <Stack
-            gap={1}
-            width="100%"
-            height="100%"
-            paddingY={6}
-            alignItems="center"
-            justifyContent="center"
-          >
-            <FileUploadOutlinedIcon sx={{ fontSize: "40px" }} />
-            {!loading && dragging && (
-              <Typography variant="body2">Drop files to upload</Typography>
-            )}
-            {!loading && !dragging && (
-              <>
+        {isMissingData ? (
+          <Stack spacing={-1} width="100%" height="100%" overflow="hidden">
+            <Skeleton height="50px" />
+            <Skeleton height="50px" />
+            <Skeleton height="50px" />
+            <Skeleton height="50px" />
+            <Skeleton height="50px" />
+            <Skeleton height="50px" />
+            <Skeleton height="50px" />
+            <Skeleton height="50px" />
+            <Skeleton height="50px" />
+            <Box
+              sx={{
+                position: "absolute",
+                top: "calc(50% - 40px)",
+                bottom: "16px",
+                left: "calc(50% - 100px)",
+              }}
+            >
+              <Stack
+                spacing={2}
+                sx={{ maxWidth: "210px", textAlign: "center" }}
+                alignItems="center"
+              >
                 <Typography variant="body2">
-                  Drag files here to upload
+                  <span style={{ fontWeight: "bold" }}>Drag and drop</span>{" "}
+                  files or click to the button to upload
                 </Typography>
-                <Typography
-                  variant="body2"
-                  sx={{ color: theme.palette.info.main }}
+                <Button
+                  startIcon={<AddIcon fontSize="small" />}
+                  size="small"
+                  variant="contained"
+                  onClick={(e) => {
+                    e.stopPropagation();
+                    e.preventDefault();
+                    inputRef.current?.click();
+                  }}
                 >
-                  or choose a file
-                </Typography>
-              </>
-            )}
-            {loading && (
-              <LinearProgress
-                color="info"
-                sx={{ height: "2px", width: "100%", maxWidth: "200px" }}
-              />
-            )}
+                  Upload
+                </Button>
+              </Stack>
+            </Box>
           </Stack>
+        ) : (
+          <>
+            {children ?? (
+              <Stack
+                gap={1}
+                width="100%"
+                height="100%"
+                paddingY={6}
+                alignItems="center"
+                justifyContent="center"
+              >
+                <FileUploadOutlinedIcon sx={{ fontSize: "40px" }} />
+                {!loading && dragging && (
+                  <Typography variant="body2">Drop files to upload</Typography>
+                )}
+                {!loading && !dragging && (
+                  <>
+                    <Typography variant="body2">
+                      Drag files here to upload
+                    </Typography>
+                    <Typography
+                      variant="body2"
+                      sx={{ color: theme.palette.info.main }}
+                    >
+                      or choose a file
+                    </Typography>
+                  </>
+                )}
+                {loading && (
+                  <LinearProgress
+                    color="info"
+                    sx={{ height: "2px", width: "100%", maxWidth: "200px" }}
+                  />
+                )}
+              </Stack>
+            )}
+          </>
         )}
-      </Box>
+        <TargetPopper open={dragging} parent={targetParent} />
+      </Stack>
     </>
   );
 };