+++ packages/react-mui-store-file-viewer/src/components/StoreDataGrid/DataGridParentStack.tsx
@@ -1,0 +1,13 @@
+import { Stack, StackProps } from "@mui/material";
+import { FC } from "react";
+import FileDropBox from "../FileDropBox";
+
+const DataGridParentStack: FC<StackProps> = ({ children, ...props }) => {
+  return (
+    <Stack {...props}>
+      <FileDropBox>{children}</FileDropBox>
+    </Stack>
+  );
+};
+
+export default DataGridParentStack;