+++ packages/react-mui-shared/CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.37] - 2023-09-26
+
+### Added
+
+- FieldsPreview
+- FieldsPreviewProps
+- FieldsPreviewSchema
+- FieldsForm
+- SaveButtonProps
+- FieldsFormProps
+- FieldsFormSchema
+- validatePassword
+- PasswordPolicy
+- PasswordPolicyChecks
+- FieldComponentProps
+- FieldControl
+- PangeaModal
+- PangeaDeleteModal
+
 ## [0.0.36] - 2023-09-26
 
 ### Added