+++ packages/react-mui-store-file-viewer/CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.8] - 2024-01-10
+
+### Added
+
+- Updated StoreDownloadFileViewer to support upload, rename and delete operations with the right callbacks provided
+- Added Delete as a support action from the file context menu
+
+## [0.0.7] - 2024-01-09
+
+### Added
+
+- File name conflict confirmation modal
+
+## [0.0.6] - 2024-01-08
+
+### Added
+
+- Upload popover support
+
 ## [0.0.1] - 2023-09-25
 
 ### Added